+++ carma_utils/sensorlib/test/test_ProxySensor.py
@@ -4,16 +4,16 @@ from unittest.mock import MagicMock
 import carla
 import numpy as np
 
-from src.objects.ProxySensor import ProxySensor
+from src.objects.CarlaSensor import CarlaSensor
 
-class ProxySensorTestCase(unittest.TestCase):
+class CarlaSensorTestCase(unittest.TestCase):
 
     def setUp(self):
         carla_sensor = MagicMock()
-        self.proxy_sensor = ProxySensor(carla_sensor)
+        self.proxy_sensor = CarlaSensor(carla_sensor)
 
     def test_get_position(self):
-        carla_sensor = self.proxy_sensor._ProxySensor__carla_sensor
+        carla_sensor = self.proxy_sensor._CarlaSensor__carla_sensor
         expected_position = carla.Location(1.0, 2.0, 3.0)
         carla_sensor.get_location.return_value = expected_position
 
@@ -23,7 +23,7 @@ class ProxySensorTestCase(unittest.TestCase):
         carla_sensor.get_location.assert_called_once()
 
     def test_get_points_per_second(self):
-        carla_sensor = self.proxy_sensor._ProxySensor__carla_sensor
+        carla_sensor = self.proxy_sensor._CarlaSensor__carla_sensor
         expected_points_per_second = 1000
         carla_sensor.points_per_second = expected_points_per_second
 
@@ -32,7 +32,7 @@ class ProxySensorTestCase(unittest.TestCase):
         self.assertEqual(points_per_second, expected_points_per_second)
 
     def test_get_rotation_frequency(self):
-        carla_sensor = self.proxy_sensor._ProxySensor__carla_sensor
+        carla_sensor = self.proxy_sensor._CarlaSensor__carla_sensor
         expected_rotation_frequency = 700
         carla_sensor.rotation_frequency = expected_rotation_frequency
 
@@ -41,7 +41,7 @@ class ProxySensorTestCase(unittest.TestCase):
         self.assertEqual(rotation_frequency, expected_rotation_frequency)
 
     def test_get_fov_angular_width(self):
-        carla_sensor = self.proxy_sensor._ProxySensor__carla_sensor
+        carla_sensor = self.proxy_sensor._CarlaSensor__carla_sensor
         expected_upper_fov = 30
         expected_lower_fov = -10
         carla_sensor.upper_fov = expected_upper_fov
